Move TransferProgress to src/
