+++ src/editors/BlobFileHandler.ts
@@ -7,11 +7,11 @@ import { BlobGetPropertiesResponse, BlockBlobClient } from '@azure/storage-blob'
 import * as fse from 'fs-extra';
 import { ProgressLocation, Uri, window } from 'vscode';
 import { ext } from "../extensionVariables";
+import { TransferProgress } from '../TransferProgress';
 import { BlobContainerTreeItem } from '../tree/blob/BlobContainerTreeItem';
 import { BlobTreeItem } from '../tree/blob/BlobTreeItem';
 import { createBlockBlobClient, getExistingProperties } from "../utils/blobUtils";
 import { Limits } from '../utils/limits';
-import { TransferProgress } from '../utils/TransferProgress';
 import { IRemoteFileHandler } from './IRemoteFileHandler';
 
 export class BlobFileHandler implements IRemoteFileHandler<BlobTreeItem> {
